+++ rootScripts/aggregator/setup_camguin.sh
@@ -6,12 +6,11 @@ export RUNNUM=1
 export NRUNS=1
 export SPLITNUM=0
 export MINIRUNNUM=-1
-export CAM_DEBUG=3
+export CAM_DEBUG=0
 export CAM_ALARM=1
 export CAM_AGGREGATE=1
 export CAM_POSTPAN=1
-export CAM_OUTPUTDIR=./
-#export CAM_OUTPUTDIR=/chafs2/work1/apar/aggRootfiles/
+export CAM_OUTPUTDIR=/chafs2/work1/apar/aggRootfiles/
 export POSTPAN_ROOTFILES=/chafs2/work1/apar/postpan-outputs/
 #export DITHERING_ROOTFILES=/chafs2/work1/apar/DitherCorrection/
 #export DITHERING_ROOTFILES_SLOPES=/chafs2/work1/apar/BMODextractor/
